requirements.txt
streamlit
plotly.express
pandas
streamlit_folium
folium

